Version 1.6
Added the ability to render the main window to a JPG on session save.
Added the ability for the Border to be hidden (useful when custom
backgrounds are used, and the border is unnecessary, ApexParse must be
restarted when this setting is changed)
Added Exit to the File Menu
Added Open Session Logs folder.

Split up the Options menu into Parser and Window options
Cleaned up the Window Opacity system, no longer relies on float.parse
which can fail on different cultures that don't use . as the decimal
separator
Added a slightly darkened background to the grid, should help
readability somewhat
Fixed Zanverse Player not appearing if the parser was reset